+++ SpatialGDK/Source/SpatialGDK/Public/SpatialConstants.h
@@ -84,6 +84,9 @@ const Worker_ComponentId POSITION_COMPONENT_ID = 54;
 const Worker_ComponentId PERSISTENCE_COMPONENT_ID = 55;
 const Worker_ComponentId INTEREST_COMPONENT_ID = 58;
 
+// This is a marker component used by the Runtime to define which entities are system entities.
+const Worker_ComponentId SYSTEM_COMPONENT_ID = 59;
+
 // This is a component on per-worker system entities.
 const Worker_ComponentId WORKER_COMPONENT_ID = 60;
 const Worker_ComponentId PLAYERIDENTITY_COMPONENT_ID = 61;
@@ -99,7 +102,7 @@ const Worker_ComponentId GDK_DEBUG_COMPONENT_ID = 9995;
 const Worker_ComponentId DEPLOYMENT_MAP_COMPONENT_ID = 9994;
 const Worker_ComponentId STARTUP_ACTOR_MANAGER_COMPONENT_ID = 9993;
 const Worker_ComponentId GSM_SHUTDOWN_COMPONENT_ID = 9992;
-const Worker_ComponentId HEARTBEAT_COMPONENT_ID = 9991;
+const Worker_ComponentId PLAYER_CONTROLLER_COMPONENT_ID = 9991;
 
 const Worker_ComponentId SERVER_AUTH_COMPONENT_SET_ID = 9900;
 const Worker_ComponentId CLIENT_AUTH_COMPONENT_SET_ID = 9901;
@@ -143,7 +146,8 @@ const Worker_ComponentId ACTOR_AUTH_TAG_COMPONENT_ID = 2001;
 const Worker_ComponentId ACTOR_NON_AUTH_TAG_COMPONENT_ID = 2002;
 const Worker_ComponentId LB_TAG_COMPONENT_ID = 2005;
 const Worker_ComponentId GDK_KNOWN_ENTITY_TAG_COMPONENT_ID = 2007;
-const Worker_ComponentId LAST_EC_COMPONENT_ID = 2008;
+const Worker_ComponentId TOMBSTONE_TAG_COMPONENT_ID = 2008;
+const Worker_ComponentId LAST_EC_COMPONENT_ID = 2009;
 
 const Schema_FieldId DEPLOYMENT_MAP_MAP_URL_ID = 1;
 const Schema_FieldId DEPLOYMENT_MAP_ACCEPTING_PLAYERS_ID = 2;
@@ -155,8 +159,7 @@ const Schema_FieldId STARTUP_ACTOR_MANAGER_CAN_BEGIN_PLAY_ID = 1;
 const Schema_FieldId ACTOR_COMPONENT_REPLICATES_ID = 1;
 const Schema_FieldId ACTOR_TEAROFF_ID = 3;
 
-const Schema_FieldId HEARTBEAT_EVENT_ID = 1;
-const Schema_FieldId HEARTBEAT_CLIENT_HAS_QUIT_ID = 1;
+const Schema_FieldId PLAYER_CONTROLLER_CLIENT_HAS_QUIT_ID = 1;
 
 const Schema_FieldId SHUTDOWN_MULTI_PROCESS_REQUEST_ID = 1;
 const Schema_FieldId SHUTDOWN_ADDITIONAL_SERVERS_EVENT_ID = 1;
@@ -341,31 +344,32 @@ const FString EMPTY_TEST_MAP_PATH = TEXT("/SpatialGDK/Maps/Empty");
 const FString DEV_LOGIN_TAG = TEXT("dev_login");
 
 // A list of components clients require on top of any generated data components in order to handle non-authoritative actors correctly.
-const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_NON_AUTH_CLIENT_INTEREST =
-	TArray<Worker_ComponentId>{ // Actor components
-								UNREAL_METADATA_COMPONENT_ID, SPAWN_DATA_COMPONENT_ID, TOMBSTONE_COMPONENT_ID, DORMANT_COMPONENT_ID,
+const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_NON_AUTH_CLIENT_INTEREST = TArray<Worker_ComponentId>{
+	// Actor components
+	UNREAL_METADATA_COMPONENT_ID, SPAWN_DATA_COMPONENT_ID, TOMBSTONE_COMPONENT_ID, TOMBSTONE_TAG_COMPONENT_ID, DORMANT_COMPONENT_ID,
 
-								// Multicast RPCs
-								MULTICAST_RPCS_COMPONENT_ID,
+	// Multicast RPCs
+	MULTICAST_RPCS_COMPONENT_ID,
 
-								// Global state components
-								DEPLOYMENT_MAP_COMPONENT_ID, STARTUP_ACTOR_MANAGER_COMPONENT_ID, GSM_SHUTDOWN_COMPONENT_ID,
+	// Global state components
+	DEPLOYMENT_MAP_COMPONENT_ID, STARTUP_ACTOR_MANAGER_COMPONENT_ID, GSM_SHUTDOWN_COMPONENT_ID,
 
-								// Debugging information
-								DEBUG_METRICS_COMPONENT_ID, SPATIAL_DEBUGGING_COMPONENT_ID,
+	// Debugging information
+	DEBUG_METRICS_COMPONENT_ID, SPATIAL_DEBUGGING_COMPONENT_ID,
 
-								// Non auth actor tag
-								ACTOR_NON_AUTH_TAG_COMPONENT_ID
-	};
+	// Non auth actor tag
+	ACTOR_NON_AUTH_TAG_COMPONENT_ID
+};
 
 // A list of components clients require on entities they are authoritative over on top of the components already checked out by the interest
 // query.
-const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_AUTH_CLIENT_INTEREST = TArray<Worker_ComponentId>{ // RPCs from the server
-																											SERVER_ENDPOINT_COMPONENT_ID,
+const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_AUTH_CLIENT_INTEREST =
+	TArray<Worker_ComponentId>{ // RPCs from the server
+								SERVER_ENDPOINT_COMPONENT_ID,
 
-																											// Actor auth tag
-																											ACTOR_AUTH_TAG_COMPONENT_ID
-};
+								// Actor tags
+								ACTOR_NON_AUTH_TAG_COMPONENT_ID, ACTOR_AUTH_TAG_COMPONENT_ID
+	};
 
 // A list of components servers require on top of any generated data and handover components in order to handle non-authoritative actors
 // correctly.
@@ -386,25 +390,26 @@ const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_NON_AUTH_SERVER_INTERES
 								// Authority intent component to handle scattered hierarchies
 								AUTHORITY_INTENT_COMPONENT_ID,
 
-								// Tags: Well known entities, and non-auth actors
-								GDK_KNOWN_ENTITY_TAG_COMPONENT_ID, ACTOR_NON_AUTH_TAG_COMPONENT_ID,
+								// Tags: Well known entities, non-auth actors, and tombstone tags
+								GDK_KNOWN_ENTITY_TAG_COMPONENT_ID, ACTOR_NON_AUTH_TAG_COMPONENT_ID, TOMBSTONE_TAG_COMPONENT_ID,
 
-								PARTITION_COMPONENT_ID
+								PLAYER_CONTROLLER_COMPONENT_ID, PARTITION_COMPONENT_ID
 	};
 
 // A list of components servers require on entities they are authoritative over on top of the components already checked out by the interest
 // query.
-const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_AUTH_SERVER_INTEREST = TArray<Worker_ComponentId>{ // RPCs from clients
-																											CLIENT_ENDPOINT_COMPONENT_ID,
+const TArray<Worker_ComponentId> REQUIRED_COMPONENTS_FOR_AUTH_SERVER_INTEREST =
+	TArray<Worker_ComponentId>{ // RPCs from clients
+								CLIENT_ENDPOINT_COMPONENT_ID,
 
-																											// Heartbeat
-																											HEARTBEAT_COMPONENT_ID,
+								// Player controller
+								PLAYER_CONTROLLER_COMPONENT_ID,
 
-																											// Auth actor tag
-																											ACTOR_AUTH_TAG_COMPONENT_ID,
+								// Actor tags
+								ACTOR_NON_AUTH_TAG_COMPONENT_ID, ACTOR_AUTH_TAG_COMPONENT_ID,
 
-																											PARTITION_COMPONENT_ID
-};
+								PARTITION_COMPONENT_ID
+	};
 
 inline bool IsEntityCompletenessComponent(Worker_ComponentId ComponentId)
 {
@@ -451,11 +456,11 @@ const TMap<Worker_ComponentId, FString> ServerAuthorityWellKnownComponents = {
 	{ MULTICAST_RPCS_COMPONENT_ID, "unreal.generated.UnrealMulticastRPCs" },
 };
 
-const TArray<FString> ClientAuthorityWellKnownSchemaImports = { "unreal/gdk/heartbeat.schema", "unreal/gdk/rpc_components.schema",
+const TArray<FString> ClientAuthorityWellKnownSchemaImports = { "unreal/gdk/player_controller.schema", "unreal/gdk/rpc_components.schema",
 																"unreal/generated/rpc_endpoints.schema" };
 
 const TMap<Worker_ComponentId, FString> ClientAuthorityWellKnownComponents = {
-	{ HEARTBEAT_COMPONENT_ID, "unreal.Heartbeat" },
+	{ PLAYER_CONTROLLER_COMPONENT_ID, "unreal.PlayerController" },
 	{ CLIENT_ENDPOINT_COMPONENT_ID, "unreal.generated.UnrealClientEndpoint" },
 };