+++ SpatialGDK/Source/SpatialGDK/Public/Utils/EntityFactory.h
@@ -41,7 +41,8 @@ public:
 
 	static inline bool IsClientAuthoritativeComponent(Worker_ComponentId ComponentId)
 	{
-		return ComponentId == SpatialConstants::CLIENT_ENDPOINT_COMPONENT_ID || ComponentId == SpatialConstants::HEARTBEAT_COMPONENT_ID;
+		return ComponentId == SpatialConstants::CLIENT_ENDPOINT_COMPONENT_ID
+			   || ComponentId == SpatialConstants::PLAYER_CONTROLLER_COMPONENT_ID;
 	}
 
 private: