+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/UNR-3761/SpatialTestCrossServerRPC/CrossServerRPCCube.h
@@ -16,9 +16,14 @@ public:
 
 	void GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& OutLifetimeProps) const override;
 
+	void RecordEntityId();
+
 	// Array storing the IDs of the servers from which this cube has successfully received a CrossServer RPC.
 	UPROPERTY(Replicated)
 	TArray<int> ReceivedCrossServerRPCS;
+
+	UPROPERTY(Replicated)
+	int64 AuthEntityId;
 
 	UFUNCTION(CrossServer, Reliable)
 	void CrossServerTestRPC(int SendingServerID);